jd01_10 complete Task A

+++ src/by/it/kameisha/jd01_10/Bean.java
@@ -2,17 +2,17 @@ package by.it.kameisha.jd01_10;
 
 public class Bean {
     static double sum(int a, int b) {
-        return (double) (a + b);
+        return a + b;
     }
 
     @Param(a = 100, b = 50)
     static double max(int a, int b) {
-        return (double) Math.max(a, b);
+        return Math.max(a, b);
     }
 
     @Param(a = 10, b = 20)
     double min(int a, int b) {
-        return (double) Math.min(a, b);
+        return Math.min(a, b);
     }
 
     @Param(a = 30, b = 10)
